% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugBrowser.R
\name{change.line}
\alias{change.line}
\alias{.change.line}
\title{This function uses debug.from.line to simulate the execution environment 
at a specific line of the script. It stores variable names and values 
in the var.env}
\usage{
.change.line(var.env, lines, proc.nodes, current.script)
}
\arguments{
\item{var.env}{the environment that stores the variables and current line
of "execution"}

\item{lines}{A vector of line numbers corresponding to lines of the script that had code}

\item{proc.nodes}{The procedure nodes, used for extracting the code on a line}

\item{current.script}{in the even of sourced scripts this will change to reflect the current}
}
\value{
nothing
}
\description{
This function uses debug.from.line to simulate the execution environment 
at a specific line of the script. It stores variable names and values 
in the var.env
}
