% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugBrowser.R
\name{clear.environment}
\alias{clear.environment}
\alias{.clear.environment}
\title{Clears an environment but keeps the line index
as that is needed even when all variables are gone}
\usage{
.clear.environment(var.env)
}
\arguments{
\item{var.env}{The environment to be cleared}
}
\value{
nothing
}
\description{
Clears an environment but keeps the line index
as that is needed even when all variables are gone
}
