% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugBrowser.R
\name{debug.browser}
\alias{debug.browser}
\title{Interactive time traveling script browser}
\usage{
debug.browser()
}
\value{
nothing
}
\description{
This function calls up an interactive debugging interface
similar to the browser() in core R. It has the functionality
to "step line-by-line" through a script that has finished 
execution.
}
\examples{
\dontrun{
debug.browser()
}
}
