% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugFunctions.R
\name{debug.error}
\alias{debug.error}
\alias{debug.warning}
\title{Debugging Errors and Warnings}
\usage{
debug.error(stack.overflow = FALSE)

debug.warning(..., all = FALSE)
}
\arguments{
\item{stack.overflow}{}

\item{all}{}
}
\value{



}
\description{
Debugging Errors and Warnings

debug.warning
}
\examples{

}
