% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugFromType.R
\name{debug.from.type}
\alias{debug.from.type}
\title{Lines where variable is a certain type}
\usage{
debug.from.type(var, type)
}
\arguments{
\item{var}{The variable to examine.}

\item{type}{The type to track.}
}
\value{
A list of all the lines where the inputted variable is
the inputted type.
}
\description{
Given a variable and a type, this function will return a list
of the lines where the inputted variable is that type.
}
\examples{
\dontrun{
debug.from.type(..., ...)
...
}
}
