% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugGadget.R
\name{debug.gadget}
\alias{debug.gadget}
\title{GUI for debugging functions}
\usage{
debug.gadget()
}
\value{
If 'done' event is observed, returns the data frames
returned by debug.from.line. If 'cancel' event is observed,
no return value.
}
\description{
GUI for debugging functions
}
\examples{
\dontrun{
debug.gadget()
}
}
