% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug.init}
\alias{debug.init}
\title{Initialization functions}
\usage{
debug.init(input.data = NA, dir = NULL)
}
\arguments{
\item{input.data}{A path to an R script, a prov.json file,
provenance from memory, or nothing.}

\item{dir}{A path to where to save the prov folder}
}
\value{
nothing
}
\description{
Intialize the package by running RDataTracker on a script
and/or parsing provenance into a useable format.
Debug.init must be run before anything else
}
\examples{
\dontrun{
debug.init() # if there RDataTracker has already run and
there is provenance in memory
debug.init("test.R")
debug.init("prov.json")
debug.init(ddg.json()) # ddg.json is an RDataTracker function
}
}
