% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugLineage.R
\name{debug.lineage}
\alias{debug.lineage}
\title{This function will return either the lines that led up to a variable's creation
or the other variables that the chosen variable was used to create}
\usage{
debug.lineage(..., forward = F)
}
\arguments{
\item{...}{Variables to find the lineage for. Can be single elements or vectors/lists.}

\item{forward}{Determines whether to look for lineage forward in the script or backward}
}
\value{
Returns one of two things. If no parameters were passed to the function
then a vector of possible variables will be returned. If variables were passed to the
function then a list of data frames is returned. Each data frame corresponds to one
of the variables. The names of the list will correspond to the variable passed
}
\description{
This function will return either the lines that led up to a variable's creation
or the other variables that the chosen variable was used to create
}
