% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug.prov}
\alias{debug.prov}
\alias{.debug.prov}
\title{Debug.init Helper}
\usage{
.debug.prov(input.prov, is.file = T)
}
\arguments{
\item{input.prov}{A prov.json compliant file from the system or a string from memory}

\item{is.file}{Logical stating whether or not input.prov needs to be read in from the system or not}
}
\description{
Debug.init Helper
}
