% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugSearch.R
\name{debug.search}
\alias{debug.search}
\alias{.debug.search}
\title{Search Stack Overflow}
\usage{
.debug.search(search.query, order = "desc", sort = "votes",
  tagged = "r")
}
\arguments{
\item{search.query}{What the user is searching for in stack overflow}

\item{order}{Determines the order parameter}

\item{sort}{Determines how to sort the results}

\item{tagged}{Determines which tags to search under}
}
\value{
a list of information converted from JSON
}
\description{
For more information see: https://api.stackexchange.com/docs/
}
\examples{
\dontrun{
debug.search("object not found")
}
}
