% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugState.R
\name{debug.state}
\alias{debug.state}
\title{The State at a Line}
\usage{
debug.state(..., script.num = 1)
}
\arguments{
\item{...}{The line numbers to be queried.}

\item{script.num}{The script number of the queried line numbers. This is ignored
if no line numbers are given.
Allows for only 1 script number to be queried per function call.
Defaults to script number 1 (main script).}
}
\value{
A list of data frames of states for each queried line number, or the state
        at the end of execution if no parameters are given to the function.
}
\description{
For each queried line, debug.state returns a data frame showing the state
at that line. This assumes that the given line has been executed.
Each data frame contains the following columns:
\itemize{
	\item name: The names of variables in the state.
	\item value: The value of each variable.
	\item container: The type of the container of each variable.
	\item dimension: The size of the container.
	\item type: The data type(s) contained within the container.
	\item scriptNum: The script number associated with each variable.
	\item startLine: The line number associated with each variable.
}
If no paramters are given, debug.state will return the state at the end of
execution.
}
\examples{
\dontrun{
prov.debug.run("test.R")
debug.state()
debug.state(5)
debug.state(10, 20, script.num = 2)
}

}
