% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugFunctions.R
\name{debug.type.changes}
\alias{debug.type.changes}
\title{Tracking Type Changes}
\usage{
debug.type.changes(var = NA)
}
\arguments{
\item{var}{Optional. Variable name(s) to be queried. If not NA, the results will
be filtered to show only those with the given variable name.}
}
\value{
A list of data frames for each variable with at least 1 data type change.
}
\description{
Returns a data frame for each variable in the execution containing the 
instances where the data type changed.
Each data frame contains the following columns:
\itemize{
	\item value: The value of the variable.
	\item container: The type of the container of the variable.
	\item dimension: The size of the container.
	\item type: The data type(s) contained within the container.
	\item code: The line of code associated with the variable.
	\item scriptNum: The script number associated with the variable.
	\item startLine: The line number associated with the variable.
}
}
\examples{
\dontrun{
prov.debug.run("test.R")
debug.type.changes()
debug.type.changes("x")
debug.type.changes(var = c("a", "b"))
}

}
