% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugFunctions.R
\name{debug.variable}
\alias{debug.variable}
\title{Tracking Changes to a Variable}
\usage{
debug.variable(..., val.type = NA, script.num = 1, all = FALSE)
}
\arguments{
\item{...}{The variable names to be queried.}

\item{val.type}{Optional. If not NA, this filters the results to contain
only instances where the valType (container or type) has the
queried type. Only one type may be queried per function call.}

\item{script.num}{The script number of the queried variables.
Defaults to script number 1 (main script).}

\item{all}{If TRUE, results for all variables of the specified script will be
returned.}
}
\value{
A list of data frames showing all instances of each variable queried.
}
\description{
For each variable queried, debug.variable returns a data frame of all
instances (data nodes) of that variable.
Each data frame contains the following columns:
\itemize{
	\item value: The value of the variable.
	\item container: The type of the container of the variable.
	\item dimension: The size of the container.
	\item type: The data type(s) contained within the container.
	\item scriptNum: The script number the variable is associated with.
	\item startLine: The line number the variable is associated with.
}
}
\examples{
\dontrun{
prov.debug.run("test.R")
debug.variable("x")
debug.variable(all = TRUE)
debug.variable("a", "b", "x", val.type = "logical")
debug.variable("a", "b", "x", script.num = 3)
}

}
