% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugVariableType.R
\name{debug.variable.type}
\alias{debug.variable.type}
\title{Type changes}
\usage{
debug.variable.type(..., just.logical = F)
}
\arguments{
\item{...}{Variables to find the type for. Can be single elements or vectors/lists.}

\item{just.logical}{Determines whether or not to display only if the variable has changed
or what the actual values were throughout execution}
}
\value{
Returns one of two things. If no parameters were passed to the function
then a vector of possible variables will be returned. If variables were passed to the
function then a list of data frames is returned. Each data frame corresponds to one
of the variables. The names of the list will correspond to the variable passed
}
\description{
This function show how a variable's type has changed throughout
execution of a script
}
\examples{
\dontrun{
debug.variable.type("x")
l <- c("x", "y", "foo", "bar")
debug.variable.type(l)
debug.variable.type(l, "z", just.logical = T)
}
}
