% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugVariableType.R
\name{grab.instances}
\alias{grab.instances}
\alias{.grab.instances}
\title{This helper function takes a variable name and finds each
instance of it in a script and what type it was. Optionally
the user can request a list that states whether or not the
variable changed type.}
\usage{
.grab.instances(result)
}
\arguments{
\item{result}{A character that corresponds to a variable}
}
\value{
A data frame that has each assignment statmenent of result as a row
}
\description{
This helper function takes a variable name and finds each
instance of it in a script and what type it was. Optionally
the user can request a list that states whether or not the
variable changed type.
}
