% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugFromLine.R
\name{grab.line}
\alias{grab.line}
\alias{.grab.line}
\title{This helper function is used to find all procedure or data nodes
on each line that the user has passed. The nodes are then processed
in another helper function which modifies the data frame line.df}
\usage{
.grab.line(lineNumber, state)
}
\arguments{
\item{lineNumber}{A numeric corresponding to the line that the
user wants to examine}

\item{state}{Determines if the variable references on that line
will be examined or the state of all variables up to that line's
execution}
}
\value{
A data frame in the debug environemnt, which contains the
columns var/code, val, type, and script. Each row is a variable.
}
\description{
This helper function is used to find all procedure or data nodes
on each line that the user has passed. The nodes are then processed
in another helper function which modifies the data frame line.df
}
