% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugLineage.R
\name{grab.lineage}
\alias{grab.lineage}
\alias{.grab.lineage}
\title{This helper function is used to find the labels of each
result the user passed. The labels are then processed in
another helper function which returns a data frame.}
\usage{
.grab.lineage(result, forward)
}
\arguments{
\item{result}{A character that corresponds to a variable name}

\item{forward}{A logical determining whether or not to search forward}
}
\value{
A data frame that contains the lineage of a variable.
Each row is a line from the script with corresponding metadata,
script #, line #, etc.
}
\description{
This helper function is used to find the labels of each
result the user passed. The labels are then processed in
another helper function which returns a data frame.
}
