% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugIssues.R
\name{process.error}
\alias{process.error}
\alias{.process.error}
\title{Process error message strings}
\usage{
.process.error(error.message)
}
\arguments{
\item{error.message}{a character vector to be cleaned}
}
\value{
character
}
\description{
This function removes *most* local information about a
script by removing all characters between quotes (single or
double, inclusive)
}
