% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugLineage.R
\name{process.label}
\alias{process.label}
\alias{.process.label}
\title{This function uses the spine of connected nodes to return the data frame
This function is also used by debug.warning.trace()}
\usage{
.process.label(label, proc.nodes, forward, assign.state = NA)
}
\arguments{
\item{label}{A character corresponding to a node name in the prov}

\item{proc.nodes}{Data frame with the prov porcedure nodes}

\item{forward}{A logical determining whether or not to search forward}

\item{assign.state}{A possibly NULL character used if lineage is going forward.
It has the label for the assignment statement for the chosen variable.}
}
\value{
A data frame that contains the lineage of a variable.
Each row is a line from the script with corresponding metadata,
script #, line #, etc.
}
\description{
This function uses the spine of connected nodes to return the data frame
This function is also used by debug.warning.trace()
}
