% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runScriptAddin.R
\name{run.script.addin}
\alias{run.script.addin}
\title{Serves as an interface with rstudioapi to facilitate use by RStudio users}
\usage{
run.script.addin(input.path = NA)
}
\arguments{
\item{input.path}{A path to a file to run. If not given, will pull a script from where the cursor is.
If the cursor has highlighted text will grab that and only run the highlighted text.}
}
\value{
nothing
}
\description{
This function is bound to a button that will show up undeer the RSudio addins menu.
For more information on addins see: https://www.rstudio.com/resources/webinars/understanding-add-ins/
Preferably that button will be bound to a keyboard shortcut. This will allow users
to hit a keyboard shortcut to run their script and collect prov
}
\examples{
\dontrun{
# This function will be run by navigating to the Addins menu in RStudio
# and then clicking the Run w/ Debugging button. It is reccommended this
# button is asssigned to a keyboard shortcut. 
}
}
