% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugGadget.R
\name{stableColumnLayout}
\alias{stableColumnLayout}
\alias{.stableColumnLayout}
\title{This helper function from the rstudio/addinexample repo formats
a stable column layout.}
\usage{
.stableColumnLayout(...)
}
\arguments{
\item{...}{UI elements}
}
\value{
A list of shiny tags, or functions that parallel common 
HTML tags, representing each UI element argument.
}
\description{
This helper function from the rstudio/addinexample repo formats
a stable column layout.
}
