% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug.init}
\alias{debug.init}
\alias{debug.prov}
\title{Intialize the package by running RDataTracker on a script
and/or parsing the resulting provenance into a useable format
Debug.init or debug.prov must be run before anything else}
\usage{
debug.init(input.data)

debug.prov(input.prov, is.file = T)
}
\arguments{
\item{input.data}{A path to an R script}

\item{input.prov}{A prov.json compliant file from the system or a string from memory}

\item{is.file}{Logical stating whether or not input.prov needs to be read in from the system or not}
}
\value{
There is no return value
}
\description{
Intialize the package by running RDataTracker on a script
and/or parsing the resulting provenance into a useable format
Debug.init or debug.prov must be run before anything else
}
\examples{
\dontrun{
debug.init("test.R")
}
\dontrun{
debug.prov("prov.json")
}
}
