% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugLine.R
\name{debug.line}
\alias{debug.line}
\title{Tracking the Inputs and Outputs at a Line}
\usage{
debug.line(..., script.num = 1, all = FALSE)
}
\arguments{
\item{...}{The line numbers to be queried.}

\item{script.num}{The script number of the queried line numbers.
Allows for only 1 script number to be queried per function call.
Defaults to script number 1 (main script).}

\item{all}{If TRUE, the inputs and outputs for all lines in the specified script
will be returned.}
}
\value{
A list of data frames showing the inputs and outputs for the procedure
        in each line queried.
}
\description{
For each line number queried, debug.line returns a data frame of the data 
that the procedure in that line inputs and outputs.
Each data frame contains the following columns:
\itemize{
	\item name: The name of the data.
	\item value: The value of the data.
	\item container: The type of the container of the data.
	\item dimension: The size of the container.
	\item type: The data type(s) contained within the container.
}
}
\examples{
\dontrun{
prov.debug.run("test.R")
debug.line(5)
debug.line(all = TRUE)
debug.line(5, 10, script.num = 2)
}

}
