% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugFromLine.R
\name{process.node}
\alias{process.node}
\alias{.process.node}
\title{This helper function is used to find information about the node
passed to it. A row with the information is created and appended
to a data frame in the debug environment.}
\usage{
.process.node(node, script.num)
}
\arguments{
\item{node}{A character corresponding to a node name in the prov}

\item{script.num}{If 1, examine only variables from the main script.
If 2, examine variables from first nested source script.}
}
\value{
Nothing
}
\description{
This helper function is used to find information about the node
passed to it. A row with the information is created and appended
to a data frame in the debug environment.
}
